chore: move all schema version crud logic from legacy storage module to dedicated class (#8029)
