+++ packages/services/api/src/modules/schema/providers/schema-manager.ts
@@ -43,9 +43,9 @@ import { TargetManager } from '../../target/providers/target-manager';
 import { BreakingSchemaChangeUsageHelper } from './breaking-schema-changes-helper';
 import { SCHEMA_MODULE_CONFIG, type SchemaModuleConfig } from './config';
 import { Contracts } from './contracts';
-import type { SchemaCoordinatesDiffResult } from './inspector';
 import { CompositionOrchestrator } from './orchestrator/composition-orchestrator';
 import { ensureCompositeSchemas, removeDescriptions, SchemaHelper } from './schema-helper';
+import { SchemaVersionStore } from './schema-version-store';
 
 const ENABLE_EXTERNAL_COMPOSITION_SCHEMA = z.object({
   endpoint: z.string().url().nonempty(),
@@ -84,6 +84,7 @@ export class SchemaManager {
     private contracts: Contracts,
     private breakingSchemaChangeUsageHelper: BreakingSchemaChangeUsageHelper,
     private idTranslator: IdTranslator,
+    private schemaVersions: SchemaVersionStore,
     @Inject(SCHEMA_MODULE_CONFIG) private schemaModuleConfig: SchemaModuleConfig,
   ) {
     this.logger = logger.child({ source: 'SchemaManager' });
@@ -92,7 +93,9 @@ export class SchemaManager {
         return Promise.all(
           selectors.map(async selector => {
             return {
-              ...(await this.storage.getLatestValidVersion(selector)),
+              ...(await this.schemaVersions.getLatestValidSchemaVersionForTargetId(
+                selector.targetId,
+              )),
               projectId: selector.projectId,
               targetId: selector.targetId,
               organizationId: selector.organizationId,
@@ -110,11 +113,7 @@ export class SchemaManager {
 
   async hasSchema(target: Target) {
     this.logger.debug('Checking if schema is available (targetId=%s)', target.id);
-    return this.storage.hasSchema({
-      organizationId: target.orgId,
-      projectId: target.projectId,
-      targetId: target.id,
-    });
+    return this.schemaVersions.anyVersionExistsForTarget(target);
   }
 
   @traceFn('SchemaManager.compose', {
@@ -255,7 +254,7 @@ export class SchemaManager {
     } & TargetSelector,
   ) {
     this.logger.debug('Fetching non-empty list of schemas (selector=%o)', selector);
-    const schemas = await this.storage.getSchemasOfVersion(selector);
+    const schemas = await this.schemaVersions.getSchemasBySchemaVersionId(selector.versionId);
 
     if (schemas.length === 0) {
       throw new HiveError('No schemas found for this version.');
@@ -275,19 +274,17 @@ export class SchemaManager {
   @atomic(stringifySelector)
   async getMaybeSchemasOfVersion(schemaVersion: SchemaVersion) {
     this.logger.debug('Fetching schemas (schemaVersionId=%s)', schemaVersion.id);
-    return this.storage.getSchemasOfVersion({ versionId: schemaVersion.id });
+    return this.schemaVersions.getSchemasBySchemaVersionId(schemaVersion.id);
   }
 
   async getMatchingServiceSchemaOfVersions(versions: { before: string | null; after: string }) {
     this.logger.debug('Fetching service schema of versions (selector=%o)', versions);
-    return this.storage.getMatchingServiceSchemaOfVersions(versions);
+    return this.schemaVersions.getMatchingServiceSchemaOfVersions(versions);
   }
 
   async getMaybeLatestValidVersion(target: Target) {
     this.logger.debug('Fetching maybe latest valid version (targetId=%o)', target.id);
-    const version = await this.storage.getMaybeLatestValidVersion({
-      targetId: target.id,
-    });
+    const version = await this.schemaVersions.getMaybeLatestValidSchemaVersion(target);
 
     if (!version) {
       return null;
@@ -308,11 +305,7 @@ export class SchemaManager {
 
   async getMaybeLatestVersion(target: Target) {
     this.logger.debug('Fetching maybe latest version (targetId=%o)', target.id);
-    const latest = await this.storage.getMaybeLatestVersion({
-      targetId: target.id,
-      projectId: target.projectId,
-      organizationId: target.orgId,
-    });
+    const latest = await this.schemaVersions.getMaybeLatestSchemaVersionForTargetId(target.id);
 
     if (!latest) {
       return null;
@@ -326,7 +319,7 @@ export class SchemaManager {
     };
   }
 
-  async getSchemaVersion(
+  async getSchemaVersionBySelector(
     selector: TargetSelector & { versionId: string },
   ): Promise<SchemaVersion | null> {
     this.logger.debug('Fetching single schema version (selector=%o)', selector);
@@ -336,17 +329,20 @@ export class SchemaManager {
       return null;
     }
 
-    const result = await this.storage.getMaybeVersion(selector);
+    const version = await this.schemaVersions.getSchemaVersionById(selector.versionId);
 
-    if (!result) {
+    if (!version) {
+      return null;
+    }
+
+    if (version.targetId !== selector.targetId) {
       return null;
     }
 
     return {
       projectId: selector.projectId,
-      targetId: selector.targetId,
       organizationId: selector.organizationId,
-      ...result,
+      ...version,
     };
   }
 
@@ -362,10 +358,7 @@ export class SchemaManager {
       return null;
     }
 
-    const schemas = await this.storage.getSchemasOfVersion({
-      versionId: schemaVersion.id,
-      includeMetadata: true,
-    });
+    const schemas = await this.schemaVersions.getSchemasBySchemaVersionId(schemaVersion.id);
 
     return {
       version: schemaVersion,
@@ -373,14 +366,14 @@ export class SchemaManager {
     };
   }
 
-  async getPaginatedSchemaVersionsForTargetId(args: {
-    targetId: string;
-    organizationId: string;
-    projectId: string;
-    first: number | null;
-    cursor: null | string;
-  }) {
-    const connection = await this.storage.getPaginatedSchemaVersionsForTargetId(args);
+  async getPaginatedSchemaVersionsForTargetId(
+    target: Target,
+    args: {
+      first: number | null;
+      cursor: null | string;
+    },
+  ) {
+    const connection = await this.schemaVersions.getPaginatedSchemaVersionsForTarget(target, args);
 
     return {
       ...connection,
@@ -388,9 +381,9 @@ export class SchemaManager {
         ...edge,
         node: {
           ...edge.node,
-          organizationId: args.organizationId,
-          projectId: args.projectId,
-          targetId: args.targetId,
+          organizationId: target.orgId,
+          projectId: target.projectId,
+          targetId: target.id,
         },
       })),
     };
@@ -411,7 +404,7 @@ export class SchemaManager {
 
   async getSchemaLog(selector: { commit: string } & TargetSelector) {
     this.logger.debug('Fetching schema log (selector=%o)', selector);
-    return this.storage.getSchemaLog({
+    return this.schemaVersions.getSchemLog({
       commit: selector.commit,
       targetId: selector.targetId,
     });
@@ -438,10 +431,8 @@ export class SchemaManager {
       url?: string | null;
       base_schema: string | null;
       metadata: string | null;
-      projectType: ProjectType;
       actionFn(versionId: string): Promise<void>;
       changes: Array<SchemaChangeType>;
-      coordinatesDiff: SchemaCoordinatesDiffResult | null;
       previousSchemaVersion: string | null;
       diffSchemaVersionId: string | null;
       github: null | {
@@ -489,7 +480,6 @@ export class SchemaManager {
         'service',
         'logIds',
         'url',
-        'projectType',
         'previousSchemaVersion',
         'diffSchemaVersionId',
         'github',
@@ -497,7 +487,7 @@ export class SchemaManager {
       ]),
     );
 
-    return this.storage.createVersion({
+    return this.schemaVersions.createSchemaVersion({
       ...input,
       logIds: input.logIds,
     });
@@ -589,22 +579,14 @@ export class SchemaManager {
     await this.storage.updateBaseSchema(selector, newBaseSchema);
   }
 
-  countSchemaVersionsOfProject(
-    selector: ProjectSelector & {
-      period: DateRange | null;
-    },
-  ): Promise<number> {
-    this.logger.debug('Fetching schema versions count of project (selector=%o)', selector);
-    return this.storage.countSchemaVersionsOfProject(selector);
+  countSchemaVersionsOfProject(project: Project, period: DateRange | null): Promise<number> {
+    this.logger.debug('Fetching schema versions count of project (projectId=%s)', project.id);
+    return this.schemaVersions.countSchemaVersionsOfProject(project, period);
   }
 
-  countSchemaVersionsOfTarget(
-    selector: TargetSelector & {
-      period: DateRange | null;
-    },
-  ): Promise<number> {
-    this.logger.debug('Fetching schema versions count of target (selector=%o)', selector);
-    return this.storage.countSchemaVersionsOfTarget(selector);
+  countSchemaVersionsOfTarget(target: Target, period: DateRange | null): Promise<number> {
+    this.logger.debug('Fetching schema versions count of target (targetId=%s)', target.id);
+    return this.schemaVersions.countSchemaVersionsOfTarget(target, period);
   }
 
   async completeGetStartedCheck(
@@ -1084,11 +1066,8 @@ export class SchemaManager {
       },
     });
 
-    const record = await this.storage.getSchemaVersionByCommit({
-      projectId: selector.projectId,
-      targetId: selector.targetId,
-      commit: args.commit,
-    });
+    const target = await this.targetManager.getTargetById({ targetId: selector.targetId });
+    const record = await this.schemaVersions.getSchemaVersionForTargetByCommit(target, args.commit);
 
     if (!record) {
       return null;
@@ -1102,47 +1081,31 @@ export class SchemaManager {
     };
   }
 
-  async getComposableVersionBeforeVersionId(args: {
-    organization: string;
-    project: string;
-    target: string;
-    beforeVersionId: string;
-    beforeVersionCreatedAt: string;
-  }) {
-    this.logger.debug('Fetch version before version id. (args=%o)', args);
+  async getComposableVersionBeforeVersionId(schemaVersion: SchemaVersion) {
+    this.logger.debug('Fetch version before version id. (schemaVersionId=%s)', schemaVersion.id);
 
-    const schemaVersion = await this.storage.getVersionBeforeVersionId({
-      targetId: args.target,
-      beforeVersionId: args.beforeVersionId,
-      beforeVersionCreatedAt: args.beforeVersionCreatedAt,
-      onlyComposable: true,
-    });
+    const previousSchemaVersion = await this.schemaVersions.getSchemaVersionBeforeSchemaVersion(
+      schemaVersion,
+      true,
+    );
 
-    if (!schemaVersion) {
+    if (!previousSchemaVersion) {
       return null;
     }
 
     return {
-      ...schemaVersion,
-      organizationId: args.organization,
-      projectId: args.project,
-      targetId: args.target,
+      ...previousSchemaVersion,
+      organizationId: schemaVersion.organizationId,
+      projectId: schemaVersion.projectId,
+      targetId: schemaVersion.targetId,
     };
   }
 
-  async getFirstComposableSchemaVersionBeforeVersionId(args: {
-    organization: string;
-    project: string;
-    target: string;
-    beforeVersionId: string;
-    beforeVersionCreatedAt: string;
-  }) {
-    const schemaVersion = await this.storage.getVersionBeforeVersionId({
-      targetId: args.target,
-      beforeVersionId: args.beforeVersionId,
-      beforeVersionCreatedAt: args.beforeVersionCreatedAt,
-      onlyComposable: true,
-    });
+  async getFirstComposableSchemaVersionBeforeSchemaVersion(previousSchemaVersion: SchemaVersion) {
+    const schemaVersion = await this.schemaVersions.getSchemaVersionBeforeSchemaVersion(
+      previousSchemaVersion,
+      true,
+    );
 
     if (!schemaVersion) {
       return null;
@@ -1150,9 +1113,9 @@ export class SchemaManager {
 
     return {
       ...schemaVersion,
-      organizationId: args.organization,
-      projectId: args.project,
-      targetId: args.target,
+      organizationId: previousSchemaVersion.organizationId,
+      projectId: previousSchemaVersion.projectId,
+      targetId: previousSchemaVersion.targetId,
     };
   }
