+++ packages/services/api/src/modules/schema/resolvers/Target.ts
@@ -2,6 +2,7 @@ import { parseDateRangeInput } from '../../../shared/helpers';
 import { OperationsManager } from '../../operations/providers/operations-manager';
 import { ContractsManager } from '../providers/contracts-manager';
 import { SchemaManager } from '../providers/schema-manager';
+import { SchemaVersionStore } from '../providers/schema-version-store';
 import { toGraphQLSchemaCheck, toGraphQLSchemaCheckCurry } from '../to-graphql-schema-check';
 import type { TargetResolvers } from './../../../__generated__/types';
 
@@ -21,32 +22,18 @@ export const Target: Pick<
   | 'schemaVersionsCount'
 > = {
   schemaVersions: async (target, args, { injector }) => {
-    return injector.get(SchemaManager).getPaginatedSchemaVersionsForTargetId({
-      targetId: target.id,
-      organizationId: target.orgId,
-      projectId: target.projectId,
+    return injector.get(SchemaManager).getPaginatedSchemaVersionsForTargetId(target, {
       cursor: args.after ?? null,
       first: args.first ?? null,
     });
   },
   schemaVersion: async (target, args, { injector }) => {
-    const schemaVersion = await injector.get(SchemaManager).getSchemaVersion({
+    return await injector.get(SchemaManager).getSchemaVersionBySelector({
       organizationId: target.orgId,
       projectId: target.projectId,
       targetId: target.id,
       versionId: args.id,
     });
-
-    if (schemaVersion === null) {
-      return null;
-    }
-
-    return {
-      ...schemaVersion,
-      organizationId: target.orgId,
-      projectId: target.projectId,
-      targetId: target.id,
-    };
   },
   latestSchemaVersion: (target, _, { injector }) => {
     return injector.get(SchemaManager).getMaybeLatestVersion(target);
@@ -58,7 +45,7 @@ export const Target: Pick<
     return injector.get(SchemaManager).getBaseSchemaForTarget(target);
   },
   hasSchema: (target, _, { injector }) => {
-    return injector.get(SchemaManager).hasSchema(target);
+    return injector.get(SchemaVersionStore).anyVersionExistsForTarget(target);
   },
   schemaCheck: async (target, args, { injector }) => {
     const schemaCheck = await injector.get(SchemaManager).findSchemaCheckForTarget(target, args.id);
@@ -92,12 +79,9 @@ export const Target: Pick<
     };
   },
   schemaVersionsCount: (target, { period }, { injector }) => {
-    return injector.get(SchemaManager).countSchemaVersionsOfTarget({
-      organizationId: target.orgId,
-      projectId: target.projectId,
-      targetId: target.id,
-      period: period ? parseDateRangeInput(period) : null,
-    });
+    return injector
+      .get(SchemaManager)
+      .countSchemaVersionsOfTarget(target, period ? parseDateRangeInput(period) : null);
   },
   contracts: async (target, args, { injector }) => {
     return await injector.get(ContractsManager).getPaginatedContractsForTarget({
