+++ packages/services/storage/src/index.ts
@@ -17,23 +17,18 @@ import {
   TaggedTemplateLiteralInvocation,
   UniqueIntegrityConstraintViolationError,
 } from '@hive/postgres';
-import type { SchemaCoordinatesDiffResult } from '../../api/src/modules/schema/providers/inspector';
 import { createSDLHash, ProjectType } from '../../api/src/shared/entities';
 import { batch, batchBy } from '../../api/src/shared/helpers';
 import { type organizations } from './db';
 import {
   AffectedAppDeployments,
-  ConditionalBreakingChangeMetadata,
-  ConditionalBreakingChangeMetadataModel,
   HiveSchemaChangeModel,
   InsertConditionalBreakingChangeMetadataModel,
   SchemaCheckModel,
-  SchemaCompositionErrorModel,
   SchemaPolicyWarningModel,
   TargetBreadcrumbModel,
   type SchemaChangeType,
   type SchemaCheckApprovalMetadata,
-  type SchemaCompositionError,
 } from './schema-change-model';
 
 export type { Interceptor };
@@ -1853,633 +1848,6 @@ export async function createStorage(
         .then(TargetSettingsModel.parse)
         .then(r => r.appDeploymentProtection);
     },
-
-    async countSchemaVersionsOfProject({ projectId: project, period }) {
-      if (period) {
-        const result = await pool
-          .maybeOne(
-            psql`/* countPeriodSchemaVersionsOfProject */
-            SELECT COUNT(*) as total FROM schema_versions as sv
-            LEFT JOIN targets as t ON (t.id = sv.target_id)
-            WHERE
-              t.project_id = ${project}
-              AND sv.created_at >= ${period.from.toISOString()}
-              AND sv.created_at < ${period.to.toISOString()}
-          `,
-          )
-          .then(z.object({ total: z.number() }).nullable().parse);
-        return result?.total ?? 0;
-      }
-
-      const result = await pool
-        .maybeOne(
-          psql`/* countSchemaVersionsOfProject */
-          SELECT COUNT(*) as total FROM schema_versions as sv
-          LEFT JOIN targets as t ON (t.id = sv.target_id)
-          WHERE t.project_id = ${project}
-        `,
-        )
-        .then(z.object({ total: z.number() }).nullable().parse);
-
-      return result?.total ?? 0;
-    },
-    async countSchemaVersionsOfTarget({ targetId: target, period }) {
-      if (period) {
-        const result = await pool
-          .maybeOne(
-            psql`/* countPeriodSchemaVersionsOfTarget */
-            SELECT COUNT(*) as total FROM schema_versions
-            WHERE
-              target_id = ${target}
-              AND created_at >= ${period.from.toISOString()}
-              AND created_at < ${period.to.toISOString()}
-          `,
-          )
-          .then(z.object({ total: z.number() }).nullable().parse);
-        return result?.total ?? 0;
-      }
-
-      const result = await pool
-        .maybeOne(
-          psql`/* countSchemaVersionsOfTarget */
-          SELECT COUNT(*) as total FROM schema_versions WHERE target_id = ${target}
-        `,
-        )
-        .then(z.object({ total: z.number() }).nullable().parse);
-
-      return result?.total ?? 0;
-    },
-
-    async hasSchema({ targetId: target }) {
-      return pool.exists(
-        psql`/* hasSchema */
-          SELECT 1 FROM schema_versions as v WHERE v.target_id = ${target} LIMIT 1
-        `,
-      );
-    },
-    async getMaybeLatestValidVersion(args) {
-      const version = await pool.maybeOne(
-        psql`/* getMaybeLatestValidVersion */
-          SELECT
-            ${schemaVersionSQLFields(psql`sv.`)}
-          FROM schema_versions as sv
-          WHERE sv.target_id = ${args.targetId} AND sv.is_composable IS TRUE
-          ORDER BY sv.created_at DESC
-          LIMIT 1
-        `,
-      );
-
-      if (!version) {
-        return null;
-      }
-
-      return SchemaVersionModel.parse(version);
-    },
-    async getLatestValidVersion({ targetId: target }) {
-      const version = await pool.maybeOne(
-        psql`/* getLatestValidVersion */
-          SELECT
-            ${schemaVersionSQLFields(psql`sv.`)}
-          FROM schema_versions as sv
-          WHERE sv.target_id = ${target} AND sv.is_composable IS TRUE
-          ORDER BY sv.created_at DESC
-          LIMIT 1
-        `,
-      );
-
-      return SchemaVersionModel.parse(version);
-    },
-    async getMaybeLatestVersion(args) {
-      const version = await pool.maybeOne(
-        psql`/* getMaybeLatestVersion */
-          SELECT
-            ${schemaVersionSQLFields(psql`sv.`)}
-          FROM
-            "schema_versions" AS "sv"
-          WHERE
-            "sv"."target_id" = ${args.targetId}
-          ORDER BY
-            "sv"."created_at" DESC
-          LIMIT 1
-        `,
-      );
-
-      if (!version) {
-        return null;
-      }
-
-      return SchemaVersionModel.parse(version);
-    },
-    async getVersionBeforeVersionId(args) {
-      const version = await pool.maybeOne(
-        psql`/* getVersionBeforeVersionId */
-          SELECT
-            ${schemaVersionSQLFields()}
-          FROM "schema_versions"
-          WHERE
-            "target_id" = ${args.targetId}
-            AND (
-              (
-                "created_at" = ${args.beforeVersionCreatedAt}
-                AND "id" < ${args.beforeVersionId}
-              )
-              OR "created_at" < ${args.beforeVersionCreatedAt}
-            )
-            ${args.onlyComposable ? psql`AND "is_composable" = TRUE` : psql``}
-          ORDER BY
-            "created_at" DESC
-          LIMIT 1
-        `,
-      );
-
-      if (!version) {
-        return null;
-      }
-
-      return SchemaVersionModel.parse(version);
-    },
-    async getSchemaByNameOfVersion(args) {
-      return pool
-        .maybeOne(
-          psql`/* getSchemaByNameOfVersion */
-            SELECT
-              ${schemaLogFields(psql`sl.`)}
-              , p.type
-            FROM schema_version_to_log AS svl
-            LEFT JOIN schema_log AS sl ON (sl.id = svl.action_id)
-            LEFT JOIN projects as p ON (p.id = sl.project_id)
-            WHERE
-              svl.version_id = ${args.versionId}
-              AND sl.action = 'PUSH'
-              AND p.type != 'CUSTOM'
-              AND lower(sl.service_name) = lower(${args.serviceName})
-            ORDER BY
-              sl.created_at DESC
-          `,
-        )
-        .then(SchemaPushLogModel.nullable().parse);
-    },
-    async getSchemasOfVersion({ versionId: version, includeMetadata: _includeMetadata = false }) {
-      return pool
-        .any(
-          psql`/* getSchemasOfVersion */
-            SELECT
-              ${schemaLogFields(psql`sl.`)}
-              , p.type
-            FROM schema_version_to_log AS svl
-            LEFT JOIN schema_log AS sl ON (sl.id = svl.action_id)
-            LEFT JOIN projects as p ON (p.id = sl.project_id)
-            WHERE
-              svl.version_id = ${version}
-              AND sl.action = 'PUSH'
-              AND p.type != 'CUSTOM'
-            ORDER BY
-              sl.created_at DESC
-          `,
-        )
-        .then(z.array(SchemaPushLogModel).parse);
-    },
-    async getServiceSchemaOfVersion(args) {
-      return pool
-        .maybeOne(
-          psql`/* getServiceSchemaOfVersion */
-          SELECT
-              ${schemaLogFields(psql`sl.`)}
-              , p.type
-            FROM schema_version_to_log AS svl
-            LEFT JOIN schema_log AS sl ON (sl.id = svl.action_id)
-            LEFT JOIN projects as p ON (p.id = sl.project_id)
-            WHERE
-              svl.version_id = ${args.schemaVersionId}
-              AND sl.action = 'PUSH'
-              AND p.type != 'CUSTOM'
-              AND lower(sl.service_name) = lower(${args.serviceName})
-        `,
-        )
-        .then(SchemaPushLogModel.nullable().parse);
-    },
-
-    async getMatchingServiceSchemaOfVersions(versions) {
-      const after = await pool
-        .maybeOne(
-          psql`/* getMatchingServiceSchemaOfVersions */
-          SELECT sl.service_name, sl.sdl
-          FROM schema_versions as sv
-          LEFT JOIN schema_log as sl ON sv.action_id = sl.id
-          WHERE sv.id = ${versions.after} AND service_name IS NOT NULL
-        `,
-        )
-        .then(z.object({ service_name: z.string(), sdl: z.string() }).parse);
-
-      // It's an initial version, so we just need to fetch a single version
-      if (!versions.before) {
-        return { serviceName: after.service_name, after: after.sdl, before: null };
-      }
-
-      const before = await pool
-        .maybeOne(
-          psql`/* getMatchingServiceSchemaOfVersions */
-          SELECT sl.sdl
-          FROM schema_version_to_log as svtl
-          LEFT JOIN schema_log as sl ON svtl.action_id = sl.id
-          WHERE svtl.version_id = ${versions.before} AND sl.service_name = ${after.service_name}
-        `,
-        )
-        .then(z.object({ sdl: z.string().nullable() }).nullable().parse);
-
-      return { serviceName: after.service_name, after: after.sdl, before: before?.sdl ?? null };
-    },
-
-    async getMaybeVersion({ projectId: project, targetId: target, versionId: version }) {
-      const result = await pool.maybeOne(psql`/* getMaybeVersion */
-        SELECT
-          ${schemaVersionSQLFields(psql`sv.`)}
-        FROM schema_versions as sv
-        LEFT JOIN schema_log as sl ON (sl.id = sv.action_id)
-        LEFT JOIN targets as t ON (t.id = sv.target_id)
-        WHERE
-          sv.target_id = ${target}
-          AND t.project_id = ${project}
-          AND sv.id = ${version}
-        LIMIT 1
-      `);
-
-      if (!result) {
-        return null;
-      }
-
-      return SchemaVersionModel.parse(result);
-    },
-    async getPaginatedSchemaVersionsForTargetId(args) {
-      let cursor: null | {
-        createdAt: string;
-        id: string;
-      } = null;
-
-      const limit = args.first ? (args.first > 0 ? Math.min(args.first, 20) : 20) : 20;
-
-      if (args.cursor) {
-        cursor = decodeCreatedAtAndUUIDIdBasedCursor(args.cursor);
-      }
-
-      const query = psql`/* getPaginatedSchemaVersionsForTargetId */
-        SELECT
-          ${schemaVersionSQLFields()}
-        FROM
-          "schema_versions"
-        WHERE
-          "target_id" = ${args.targetId}
-          ${
-            cursor
-              ? psql`
-                AND (
-                  (
-                    "created_at" = ${cursor.createdAt}
-                    AND "id" < ${cursor.id}
-                  )
-                  OR "created_at" < ${cursor.createdAt}
-                )
-              `
-              : psql``
-          }
-        ORDER BY
-          "created_at" DESC
-          , "id" DESC
-        LIMIT ${limit + 1}
-      `;
-
-      const result = await pool.any(query);
-
-      let edges = result.map(row => {
-        const node = SchemaVersionModel.parse(row);
-
-        return {
-          node,
-          get cursor() {
-            return encodeCreatedAtAndUUIDIdBasedCursor(node);
-          },
-        };
-      });
-
-      const hasNextPage = edges.length > limit;
-      edges = edges.slice(0, limit);
-
-      return {
-        edges,
-        pageInfo: {
-          hasNextPage,
-          hasPreviousPage: cursor !== null,
-          get endCursor() {
-            return edges[edges.length - 1]?.cursor ?? '';
-          },
-          get startCursor() {
-            return edges[0]?.cursor ?? '';
-          },
-        },
-      };
-    },
-    async deleteSchema(args) {
-      return pool.transaction('deleteSchema', async trx => {
-        // fetch the latest version
-        const latestVersion = await trx
-          .maybeOne(
-            psql`/* findLatestSchemaVersion */
-            SELECT sv.id, sv.base_schema
-            FROM schema_versions as sv
-            WHERE sv.target_id = ${args.targetId}
-            ORDER BY sv.created_at DESC
-            LIMIT 1
-          `,
-          )
-          .then(z.object({ id: z.string(), base_schema: z.string().nullable() }).parse);
-
-        // create a new action
-        const deleteActionResult = await trx
-          .maybeOne(
-            psql`/* createDeleteActionSchemaLog */
-            INSERT INTO schema_log
-              (
-                author,
-                commit,
-                service_name,
-                project_id,
-                target_id,
-                action
-              )
-            VALUES
-              (
-                ${'system'}::text,
-                ${'system'}::text,
-                lower(${args.serviceName}::text),
-                ${args.projectId},
-                ${args.targetId},
-                'DELETE'
-              )
-            RETURNING
-              id
-              , to_json("created_at") AS "createdAt"
-              , "service_name" AS "serviceName"
-              , "target_id" AS "targetId"
-          `,
-          )
-          .then(
-            z.object({
-              id: z.string(),
-              createdAt: z.string(),
-              serviceName: z.string(),
-              targetId: z.string(),
-            }).parse,
-          );
-
-        // creates a new version
-        const newVersion = await insertSchemaVersion(trx, {
-          isComposable: args.composable,
-          targetId: args.targetId,
-          actionId: deleteActionResult.id,
-          baseSchema: latestVersion.base_schema,
-          previousSchemaVersion: latestVersion.id,
-          diffSchemaVersionId: args.diffSchemaVersionId,
-          compositeSchemaSDL: args.compositeSchemaSDL,
-          supergraphSDL: args.supergraphSDL,
-          schemaCompositionErrors: args.schemaCompositionErrors,
-          // Deleting a schema is done via CLI and not associated to a commit or a pull request.
-          github: null,
-          tags: args.tags,
-          schemaMetadata: args.schemaMetadata,
-          metadataAttributes: args.metadataAttributes,
-          hasContractCompositionErrors:
-            args.contracts?.some(c => c.schemaCompositionErrors != null) ?? false,
-          conditionalBreakingChangeMetadata: args.conditionalBreakingChangeMetadata,
-        });
-
-        // Move all the schema_version_to_log entries of the previous version to the new version
-        await trx.query(psql`/* moveSchemaVersionToLog */
-          INSERT INTO schema_version_to_log
-            (version_id, action_id)
-          SELECT ${newVersion.id}::uuid as version_id, svl.action_id
-          FROM schema_version_to_log svl
-          LEFT JOIN schema_log sl ON (sl.id = svl.action_id)
-          WHERE svl.version_id = ${latestVersion.id} AND sl.action = 'PUSH' AND lower(sl.service_name) != lower(${args.serviceName})
-        `);
-
-        await trx.query(psql`/* insertSchemaVersionToLog */
-          INSERT INTO schema_version_to_log
-            (version_id, action_id)
-          VALUES
-            (${newVersion.id}, ${deleteActionResult.id})
-        `);
-
-        if (args.changes != null) {
-          await insertSchemaVersionChanges(trx, {
-            versionId: newVersion.id,
-            changes: args.changes,
-          });
-        }
-
-        if (args.coordinatesDiff) {
-          await updateSchemaCoordinateStatus(trx, {
-            targetId: args.targetId,
-            versionId: newVersion.id,
-            coordinatesDiff: args.coordinatesDiff,
-          });
-        }
-
-        for (const contract of args.contracts ?? []) {
-          const schemaVersionContractId = await insertSchemaVersionContract(trx, {
-            schemaVersionId: newVersion.id,
-            contractId: contract.contractId,
-            contractName: contract.contractName,
-            schemaCompositionErrors: contract.schemaCompositionErrors,
-            compositeSchemaSDL: contract.compositeSchemaSDL,
-            supergraphSDL: contract.supergraphSDL,
-          });
-          await insertSchemaVersionContractChanges(trx, {
-            schemaVersionContractId,
-            changes: contract.changes,
-          });
-        }
-
-        await args.actionFn(newVersion.id);
-
-        return {
-          kind: 'composite',
-          id: deleteActionResult.id,
-          date: deleteActionResult.createdAt as any,
-          service_name: deleteActionResult.serviceName,
-          target: deleteActionResult.targetId,
-          action: 'DELETE',
-          versionId: newVersion.id,
-        } satisfies CompositeDeletedSchemaLog & {
-          versionId: string;
-        };
-      });
-    },
-    async createVersion(input) {
-      const url = input.url ?? null;
-      const service = input.service ?? null;
-
-      const output = await pool.transaction('createVersion', async trx => {
-        const log = await trx
-          .maybeOne(
-            psql`/* createVersion */
-            INSERT INTO schema_log
-              (
-                author,
-                service_name,
-                service_url,
-                commit,
-                sdl,
-                project_id,
-                target_id,
-                metadata,
-                action
-              )
-            VALUES
-              (
-                ${input.author},
-                lower(${service}::text),
-                ${url}::text,
-                ${input.commit}::text,
-                ${input.schema}::text,
-                ${input.projectId},
-                ${input.targetId},
-                ${input.metadata},
-                'PUSH'
-              )
-            RETURNING id
-          `,
-          )
-          .then(z.object({ id: z.string() }).parse);
-
-        // creates a new version
-        const version = await insertSchemaVersion(trx, {
-          isComposable: input.valid,
-          targetId: input.targetId,
-          actionId: log.id,
-          baseSchema: input.base_schema,
-          previousSchemaVersion: input.previousSchemaVersion,
-          diffSchemaVersionId: input.diffSchemaVersionId,
-          compositeSchemaSDL: input.compositeSchemaSDL,
-          supergraphSDL: input.supergraphSDL,
-          schemaCompositionErrors: input.schemaCompositionErrors,
-          github: input.github,
-          tags: input.tags,
-          schemaMetadata: input.schemaMetadata,
-          metadataAttributes: input.metadataAttributes,
-          hasContractCompositionErrors:
-            input.contracts?.some(c => c.schemaCompositionErrors != null) ?? false,
-          conditionalBreakingChangeMetadata: input.conditionalBreakingChangeMetadata,
-        });
-
-        await trx.query(psql`/* insertSchemaVersionToLog */
-          INSERT INTO schema_version_to_log
-            (version_id, action_id)
-          SELECT * FROM
-            ${psql.unnest(
-              input.logIds.concat(log.id).map(actionId =>
-                // Note: change.criticality.level is actually a computed value from meta
-                [version.id, actionId],
-              ),
-              ['uuid', 'uuid'],
-            )}
-        `);
-
-        await insertSchemaVersionChanges(trx, {
-          versionId: version.id,
-          changes: input.changes,
-        });
-
-        for (const contract of input.contracts ?? []) {
-          const schemaVersionContractId = await insertSchemaVersionContract(trx, {
-            schemaVersionId: version.id,
-            contractId: contract.contractId,
-            contractName: contract.contractName,
-            schemaCompositionErrors: contract.schemaCompositionErrors,
-            compositeSchemaSDL: contract.compositeSchemaSDL,
-            supergraphSDL: contract.supergraphSDL,
-          });
-          await insertSchemaVersionContractChanges(trx, {
-            schemaVersionContractId,
-            changes: contract.changes,
-          });
-        }
-
-        if (input.coordinatesDiff) {
-          await updateSchemaCoordinateStatus(trx, {
-            targetId: input.targetId,
-            versionId: version.id,
-            coordinatesDiff: input.coordinatesDiff,
-          });
-        }
-
-        await input.actionFn(version.id);
-
-        return {
-          version,
-          log,
-        };
-      });
-
-      return output.version;
-    },
-
-    async getSchemaChangesForVersion(args) {
-      // TODO: should this be paginated?
-      const changes = await pool
-        .any(
-          psql`/* getSchemaChangesForVersion */
-        SELECT
-          "change_type" as "type",
-          "meta",
-          "severity_level" as "severityLevel",
-          "is_safe_based_on_usage" as "isSafeBasedOnUsage"
-        FROM
-          "schema_version_changes"
-        WHERE
-          "schema_version_id" = ${args.versionId}
-      `,
-        )
-        .then(z.array(HiveSchemaChangeModel).parse);
-
-      if (changes.length === 0) {
-        return null;
-      }
-
-      return changes;
-    },
-
-    getSchemaLog: batch(async selectors => {
-      const rows = await pool.any(
-        psql`/* getSchemaLog */
-            SELECT
-              ${schemaLogFields(psql`sl.`)}
-              , p.type
-            FROM schema_log as sl
-            LEFT JOIN projects as p ON (p.id = sl.project_id)
-            WHERE (sl.id, sl.target_id) IN ((${psql.join(
-              selectors.map(s => psql`${s.commit}, ${s.targetId}`),
-              psql.fragment`), (`,
-            )}))
-        `,
-      );
-      const schemas = z.array(SchemaLogModel).parse(rows);
-
-      return selectors.map(selector => {
-        const schema = schemas.find(
-          row => row.id === selector.commit && row.target === selector.targetId,
-        );
-
-        if (schema) {
-          return Promise.resolve(schema);
-        }
-
-        return Promise.reject(
-          new Error(
-            `Schema log not found (commit=${selector.commit}, target=${selector.targetId})`,
-          ),
-        );
-      });
-    }),
     async addSlackIntegration({ organizationId: organization, token }) {
       await pool.any(
         psql`/* addSlackIntegration */
@@ -3741,15 +3109,15 @@ export async function createStorage(
               ${schemaSDLHash}
             , ${args.serviceName}
             , ${args.serviceUrl}
-            , ${jsonify(args.meta)}
+            , ${psql.jsonbOrNull(args.meta)}
             , ${args.targetId}
             , ${args.schemaVersionId}
             , ${args.isSuccess}
-            , ${jsonify(args.schemaCompositionErrors)}
-            , ${jsonify(args.breakingSchemaChanges?.map(toSerializableSchemaChange))}
-            , ${jsonify(args.safeSchemaChanges?.map(toSerializableSchemaChange))}
-            , ${jsonify(args.schemaPolicyWarnings?.map(w => SchemaPolicyWarningModel.parse(w)))}
-            , ${jsonify(args.schemaPolicyErrors?.map(w => SchemaPolicyWarningModel.parse(w)))}
+            , ${psql.jsonbOrNull(args.schemaCompositionErrors)}
+            , ${psql.jsonbOrNull(args.breakingSchemaChanges?.map(toSerializableSchemaChange))}
+            , ${psql.jsonbOrNull(args.safeSchemaChanges?.map(toSerializableSchemaChange))}
+            , ${psql.jsonbOrNull(args.schemaPolicyWarnings?.map(w => SchemaPolicyWarningModel.parse(w)))}
+            , ${psql.jsonbOrNull(args.schemaPolicyErrors?.map(w => SchemaPolicyWarningModel.parse(w)))}
             , ${compositeSchemaSDLHash}
             , ${supergraphSDLHash}
             , ${args.isManuallyApproved}
@@ -3764,9 +3132,9 @@ export async function createStorage(
                 c => c.breakingSchemaChanges?.length || c.safeSchemaChanges?.length,
               ) ?? false
             }
-            , ${jsonify(InsertConditionalBreakingChangeMetadataModel.parse(args.conditionalBreakingChangeMetadata))}
+            , ${psql.jsonbOrNull(InsertConditionalBreakingChangeMetadataModel.parse(args.conditionalBreakingChangeMetadata))}
             , ${args.schemaProposalId ?? null}
-            , ${jsonify(args.schemaProposalChanges?.map(toSerializableSchemaChange))}
+            , ${psql.jsonbOrNull(args.schemaProposalChanges?.map(toSerializableSchemaChange))}
           )
           RETURNING
             "id"
@@ -3808,9 +3176,9 @@ export async function createStorage(
                 , ${contract.contractId}
                 , ${compositeSchemaSdlHash}
                 , ${supergraphSchemaSdlHash}
-                , ${jsonify(contract.schemaCompositionErrors)}
-                , ${jsonify(contract.breakingSchemaChanges?.map(toSerializableSchemaChange))}
-                , ${jsonify(contract.safeSchemaChanges?.map(toSerializableSchemaChange))}
+                , ${psql.jsonbOrNull(contract.schemaCompositionErrors)}
+                , ${psql.jsonbOrNull(contract.breakingSchemaChanges?.map(toSerializableSchemaChange))}
+                , ${psql.jsonbOrNull(contract.safeSchemaChanges?.map(toSerializableSchemaChange))}
               )
             `);
           }
@@ -4460,34 +3828,6 @@ export async function createStorage(
       });
     },
 
-    async getSchemaVersionByCommit(args) {
-      const record = await pool.maybeOne(psql`/* getSchemaVersionByCommit */
-        SELECT
-          ${schemaVersionSQLFields()}
-        FROM
-          "schema_versions"
-        WHERE
-          "action_id" = (
-            SELECT
-              "id"
-            FROM
-              "schema_log"
-            WHERE
-              "schema_log"."project_id" = ${args.projectId}
-              AND "schema_log"."target_id" = ${args.targetId}
-              AND "schema_log"."commit" = ${args.commit}
-            ORDER BY "schema_log"."created_at" DESC
-            LIMIT 1
-          )
-        LIMIT 1
-      `);
-
-      if (!record) {
-        return null;
-      }
-
-      return SchemaVersionModel.parse(record);
-    },
     // Zendesk
     async setZendeskOrganizationId({ organizationId, zendeskId }) {
       await pool.query(psql`/* setZendeskOrganizationId */
@@ -4744,63 +4084,6 @@ const FeatureFlagsModel = z
       },
   );
 
-/**  This version introduced the "diffSchemaVersionId" column. */
-const SchemaVersionRecordVersion_2024_01_10_Model = z.literal('2024-01-10');
-
-const SchemaVersionRecordVersionModel = SchemaVersionRecordVersion_2024_01_10_Model;
-
-const SchemaMetadataModel = z.object({
-  name: z.string(),
-  content: z.string(),
-  source: z.nullable(z.string()).default(null),
-});
-
-const SchemaVersionModel = z.intersection(
-  z.object({
-    id: z.string(),
-    isComposable: z.boolean(),
-    createdAt: z.string(),
-    baseSchema: z.nullable(z.string()),
-    actionId: z.string(),
-    hasPersistedSchemaChanges: z.nullable(z.boolean()).transform(val => val ?? false),
-    previousSchemaVersionId: z.nullable(z.string()),
-    diffSchemaVersionId: z.nullable(z.string()),
-    compositeSchemaSDL: z.nullable(z.string()),
-    supergraphSDL: z.nullable(z.string()),
-    schemaCompositionErrors: z.nullable(z.array(SchemaCompositionErrorModel)),
-    recordVersion: z.nullable(SchemaVersionRecordVersionModel),
-    tags: z.nullable(z.array(z.string())),
-    schemaMetadata: z.nullable(z.record(z.string(), z.array(SchemaMetadataModel))),
-    metadataAttributes: z.nullable(z.record(z.string(), z.array(z.string()))),
-    hasContractCompositionErrors: z
-      .boolean()
-      .nullable()
-      .transform(val => val ?? false),
-    conditionalBreakingChangeMetadata: ConditionalBreakingChangeMetadataModel.nullable(),
-  }),
-  z
-    .union([
-      z.object({
-        githubRepository: z.string(),
-        githubSha: z.string(),
-      }),
-      z.object({
-        githubRepository: z.null(),
-        githubSha: z.null(),
-      }),
-    ])
-    .transform(val => ({
-      github: val.githubRepository
-        ? {
-            repository: val.githubRepository,
-            sha: val.githubSha,
-          }
-        : null,
-    })),
-);
-
-export type SchemaVersion = z.infer<typeof SchemaVersionModel>;
-
 const DocumentCollectionModel = z.object({
   id: z.string(),
   title: z.string(),
@@ -4822,280 +4105,6 @@ const DocumentCollectionDocumentModel = z.object({
   createdAt: z.string(),
   updatedAt: z.string(),
 });
-
-/**
- * Insert a schema version changes into the database.
- */
-async function insertSchemaVersionContractChanges(
-  trx: CommonQueryMethods,
-  args: {
-    changes: Array<SchemaChangeType> | null;
-    schemaVersionContractId: string;
-  },
-) {
-  if (!args.changes?.length) {
-    return;
-  }
-
-  await trx.query(psql`/* insertSchemaVersionContractChanges */
-    INSERT INTO "contract_version_changes" (
-      "contract_version_id",
-      "change_type",
-      "severity_level",
-      "meta",
-      "is_safe_based_on_usage"
-    )
-    SELECT * FROM
-    ${psql.unnest(
-      args.changes.map(change =>
-        // Note: change.criticality.level is actually a computed value from meta
-        [
-          args.schemaVersionContractId,
-          change.type,
-          change.criticality,
-          JSON.stringify(change.meta),
-          change.isSafeBasedOnUsage ?? false,
-        ],
-      ),
-      ['uuid', 'text', 'text', 'jsonb', 'bool'],
-    )}
-  `);
-}
-
-/**
- * Insert a schema version changes into the database.
- */
-async function insertSchemaVersionChanges(
-  trx: CommonQueryMethods,
-  args: {
-    changes: Array<SchemaChangeType>;
-    versionId: string;
-  },
-) {
-  if (args.changes.length === 0) {
-    return;
-  }
-
-  await trx.query(psql`/* insertSchemaVersionChanges */
-    INSERT INTO schema_version_changes (
-      "schema_version_id",
-      "change_type",
-      "severity_level",
-      "meta",
-      "is_safe_based_on_usage"
-    )
-    SELECT * FROM
-    ${psql.unnest(
-      args.changes.map(change =>
-        // Note: change.criticality.level is actually a computed value from meta
-        [
-          args.versionId,
-          change.type,
-          change.criticality,
-          JSON.stringify(change.meta),
-          change.isSafeBasedOnUsage ?? false,
-        ],
-      ),
-      ['uuid', 'text', 'text', 'jsonb', 'bool'],
-    )}
-  `);
-}
-
-/**
- * Insert a new schema version into the database.
- */
-async function insertSchemaVersion(
-  trx: CommonQueryMethods,
-  args: {
-    isComposable: boolean;
-    targetId: string;
-    actionId: string;
-    baseSchema: string | null;
-    previousSchemaVersion: string | null;
-    diffSchemaVersionId: string | null;
-    compositeSchemaSDL: string | null;
-    supergraphSDL: string | null;
-    schemaCompositionErrors: Array<SchemaCompositionError> | null;
-    tags: Array<string> | null;
-    schemaMetadata: Record<
-      string,
-      Array<{ name: string; content: string; source: string | null }>
-    > | null;
-    metadataAttributes: Record<string, string[]> | null;
-    hasContractCompositionErrors: boolean;
-    github: null | {
-      sha: string;
-      repository: string;
-    };
-    conditionalBreakingChangeMetadata: null | ConditionalBreakingChangeMetadata;
-  },
-) {
-  const query = psql`/* insertSchemaVersion */
-    INSERT INTO schema_versions
-      (
-        record_version,
-        is_composable,
-        target_id,
-        action_id,
-        base_schema,
-        has_persisted_schema_changes,
-        previous_schema_version_id,
-        diff_schema_version_id,
-        composite_schema_sdl,
-        supergraph_sdl,
-        schema_composition_errors,
-        github_repository,
-        github_sha,
-        tags,
-        has_contract_composition_errors,
-        conditional_breaking_change_metadata,
-        schema_metadata,
-        metadata_attributes
-      )
-    VALUES
-      (
-        '2024-01-10',
-        ${args.isComposable},
-        ${args.targetId},
-        ${args.actionId},
-        ${args.baseSchema},
-        ${true},
-        ${args.previousSchemaVersion},
-        ${args.diffSchemaVersionId},
-        ${args.compositeSchemaSDL},
-        ${args.supergraphSDL},
-        ${jsonify(args.schemaCompositionErrors)},
-        ${args.github?.repository ?? null},
-        ${args.github?.sha ?? null},
-        ${Array.isArray(args.tags) ? psql.array(args.tags, 'text') : null},
-        ${args.hasContractCompositionErrors},
-        ${jsonify(InsertConditionalBreakingChangeMetadataModel.parse(args.conditionalBreakingChangeMetadata))},
-        ${jsonify(args.schemaMetadata)},
-        ${jsonify(args.metadataAttributes)}
-      )
-    RETURNING
-      ${schemaVersionSQLFields()}
-  `;
-
-  return await trx.maybeOne(query).then(SchemaVersionModel.parse);
-}
-
-async function insertSchemaVersionContract(
-  trx: CommonQueryMethods,
-  args: {
-    schemaVersionId: string;
-    contractId: string;
-    contractName: string;
-    compositeSchemaSDL: string | null;
-    supergraphSDL: string | null;
-    schemaCompositionErrors: Array<SchemaCompositionError> | null;
-  },
-): Promise<string> {
-  const id = await trx.oneFirst(psql`/* insertSchemaVersionContract */
-    INSERT INTO "contract_versions" (
-      "schema_version_id"
-      , "contract_id"
-      , "contract_name"
-      , "schema_composition_errors"
-      , "composite_schema_sdl"
-      , "supergraph_sdl"
-    )
-    VALUES (
-      ${args.schemaVersionId}
-      , ${args.contractId}
-      , ${args.contractName}
-      , ${jsonify(args.schemaCompositionErrors)}
-      , ${args.compositeSchemaSDL}
-      , ${args.supergraphSDL}
-    )
-    RETURNING
-      "id"
-  `);
-
-  return z.string().parse(id);
-}
-
-async function updateSchemaCoordinateStatus(
-  trx: CommonQueryMethods,
-  args: {
-    targetId: string;
-    versionId: string;
-    coordinatesDiff: SchemaCoordinatesDiffResult;
-  },
-) {
-  const actions: Promise<unknown>[] = [];
-
-  if (args.coordinatesDiff.deleted) {
-    // actions.push(
-    //   trx.query(sql`/* schema_coordinate_status_deleted */
-    //   DELETE FROM schema_coordinate_status
-    //   WHERE
-    //     target_id = ${args.targetId}
-    //     AND
-    //     coordinate = ANY(${sql.array(Array.from(args.coordinatesDiff.deleted), 'text')})
-    //     AND
-    //     created_at <= NOW()
-    // `),
-    // );
-  }
-
-  if (args.coordinatesDiff.added) {
-    // actions.push(
-    //   trx.query(sql`/* schema_coordinate_status_inserted */
-    //     INSERT INTO schema_coordinate_status
-    //     ( target_id, coordinate, created_in_version_id, deprecated_at, deprecated_in_version_id )
-    //     SELECT * FROM ${sql.unnest(
-    //       Array.from(args.coordinatesDiff.added).map(coordinate => {
-    //         const isDeprecatedAsWell = args.coordinatesDiff.deprecated.has(coordinate);
-    //         return [
-    //           args.targetId,
-    //           coordinate,
-    //           args.versionId,
-    //           // if it's added and deprecated at the same time
-    //           isDeprecatedAsWell ? 'NOW()' : null,
-    //           isDeprecatedAsWell ? args.versionId : null,
-    //         ];
-    //       }),
-    //       ['uuid', 'text', 'uuid', 'date', 'uuid'],
-    //     )}
-    //   `),
-    // );
-  }
-
-  if (args.coordinatesDiff.undeprecated) {
-    // actions.push(
-    //   trx.query(sql`/* schema_coordinate_status_undeprecated */
-    //   UPDATE schema_coordinate_status
-    //   SET deprecated_at = NULL, deprecated_in_version_id = NULL
-    //   WHERE
-    //     target_id = ${args.targetId}
-    //     AND
-    //     coordinate = ANY(${sql.array(Array.from(args.coordinatesDiff.undeprecated), 'text')})
-    // `),
-    // );
-  }
-
-  await Promise.all(actions);
-
-  if (args.coordinatesDiff.deprecated) {
-    // await trx.query(sql`/* schema_coordinate_status_deprecated */
-    //   UPDATE schema_coordinate_status
-    //   SET deprecated_at = NOW(), deprecated_in_version_id = ${args.versionId}
-    //   WHERE
-    //     target_id = ${args.targetId}
-    //     AND
-    //     coordinate = ANY(${sql.array(Array.from(args.coordinatesDiff.deprecated), 'text')})
-    // `);
-  }
-}
-
-/**
- * Small helper utility for jsonifying a nullable object.
- */
-function jsonify<T>(obj: T | null | undefined) {
-  if (obj == null) return null;
-  return psql`${JSON.stringify(obj)}::jsonb`;
-}
 
 /**
  * Utility function for stripping a schema change of its computable properties for efficient storage in the database.
@@ -5163,28 +4172,6 @@ const schemaCheckSQLFields = psql`
   , c."conditional_breaking_change_metadata" as "conditionalBreakingChangeMetadata"
   , c."schema_proposal_id" as "schemaProposalId"
   , c."schema_proposal_changes" as "schemaProposalChanges"
-`;
-
-const schemaVersionSQLFields = (t = psql``) => psql`
-  ${t}"id"
-  , ${t}"is_composable" as "isComposable"
-  , to_json(${t}"created_at") as "createdAt"
-  , ${t}"action_id" as "actionId"
-  , ${t}"base_schema" as "baseSchema"
-  , ${t}"has_persisted_schema_changes" as "hasPersistedSchemaChanges"
-  , ${t}"previous_schema_version_id" as "previousSchemaVersionId"
-  , ${t}"composite_schema_sdl" as "compositeSchemaSDL"
-  , ${t}"supergraph_sdl" as "supergraphSDL"
-  , ${t}"schema_composition_errors" as "schemaCompositionErrors"
-  , ${t}"github_repository" as "githubRepository"
-  , ${t}"github_sha" as "githubSha"
-  , ${t}"diff_schema_version_id" as "diffSchemaVersionId"
-  , ${t}"record_version" as "recordVersion"
-  , ${t}"tags"
-  , ${t}"has_contract_composition_errors" as "hasContractCompositionErrors"
-  , ${t}"conditional_breaking_change_metadata" as "conditionalBreakingChangeMetadata"
-  , ${t}"schema_metadata" as "schemaMetadata"
-  , ${t}"metadata_attributes" as "metadataAttributes"
 `;
 
 const targetSQLFields = psql`
@@ -5275,19 +4262,6 @@ export {
   buildRegistryServiceURLFromMeta,
   type RegistryServiceUrlChangeSerializableChange,
 } from './schema-change-meta';
-
-export type PaginatedSchemaVersionConnection = Readonly<{
-  edges: ReadonlyArray<{
-    cursor: string;
-    node: SchemaVersion;
-  }>;
-  pageInfo: Readonly<{
-    hasNextPage: boolean;
-    hasPreviousPage: boolean;
-    startCursor: string;
-    endCursor: string;
-  }>;
-}>;
 
 export type PaginatedOrganizationInvitationConnection = Readonly<{
   edges: ReadonlyArray<{
@@ -5432,19 +4406,6 @@ const targetSettingsFields = (prefix: TaggedTemplateLiteralInvocation) => psql`
   , ${prefix}"app_deployment_protection_rule_logic" AS "appDeploymentProtectionRuleLogic"
 `;
 
-const schemaLogFields = (prefix: TaggedTemplateLiteralInvocation) => psql`
-  ${prefix}"id"
-  , ${prefix}"author"
-  , ${prefix}"commit"
-  , ${prefix}"sdl"
-  , ${prefix}"created_at" AS "date"
-  , ${prefix}"target_id" AS "target"
-  , ${prefix}"metadata"
-  , lower(${prefix}"service_name") AS "service_name"
-  , ${prefix}"service_url"
-  , ${prefix}"action"
-`;
-
 const cdnAccessTokenFields = (prefix: TaggedTemplateLiteralInvocation = psql``) => psql`
   ${prefix}"id" AS "id"
   , ${prefix}"target_id" AS "targetId"
@@ -5471,58 +4432,6 @@ const oidcIntegrationFields = (prefix: TaggedTemplateLiteralInvocation = psql``)
   , ${prefix}"default_assigned_resources" AS "defaultAssignedResources"
   , ${prefix}"require_invitation" AS "requireInvitation"
 `;
-
-const SchemaLogBase = z.object({
-  id: z.string(),
-  date: z.number(),
-  target: z.string().uuid(),
-});
-
-const SchemaPushLogBase = SchemaLogBase.extend({
-  author: z.string(),
-  commit: z.string(),
-  sdl: z.string(),
-  metadata: z.string().nullish().default(null),
-});
-
-const SinglePushSchemaLogModel = SchemaPushLogBase.extend({
-  action: z.literal('PUSH'),
-  type: z.literal('SINGLE'),
-  service_name: z.string().nullable(),
-  service_url: z.string().nullable(),
-}).transform(value => ({
-  ...value,
-  kind: 'single' as const,
-}));
-
-export type SinglePushSchemaLog = z.TypeOf<typeof SinglePushSchemaLogModel>;
-
-const CompositeSchemaTypeModel = z.union([z.literal('FEDERATION'), z.literal('STITCHING')]);
-
-const CompositePushSchemaLogModel = SchemaPushLogBase.extend({
-  action: z.literal('PUSH'),
-  type: CompositeSchemaTypeModel,
-  service_name: z.string(),
-  service_url: z.string(),
-}).transform(value => ({
-  ...value,
-  kind: 'composite' as const,
-}));
-
-export type CompositePushSchemaLog = z.TypeOf<typeof CompositePushSchemaLogModel>;
-
-const CompositeDeletedSchemaLogModel = SchemaLogBase.extend({
-  action: z.literal('DELETE'),
-  service_name: z.string(),
-}).transform(value => ({
-  ...value,
-  kind: 'composite' as const,
-}));
-
-export type CompositeDeletedSchemaLog = z.TypeOf<typeof CompositeDeletedSchemaLogModel>;
-
-const SchemaPushLogModel = z.union([SinglePushSchemaLogModel, CompositePushSchemaLogModel]);
-const SchemaLogModel = z.union([SchemaPushLogModel, CompositeDeletedSchemaLogModel]);
 
 const OrganizationModel = z
   .object({
